requirements.txt
prisma==0.13.1
fastapi==0.111.0
uvicorn==0.30.1
pytest==8.2.2
requests==2.32.3
bcrypt==3.2.2
passlib==1.7.4
python-jose==3.3.0
